scripts/chapter5/pyproject.toml
[project]
name = "chapter5-rag"
version = "0.1.0"
description = "第5章 RAG実装のサンプルコード"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain==1.0.3",
    "langchain-classic>=0.1.0",
    "langchain-openai==1.0.1",
    "langchain-community==0.4.1",
    "langchain-text-splitters==1.0.0",
    "chromadb==1.3.0",
    "matplotlib==3.10.7",
    "openai==2.6.1",
    "python-dotenv==1.2.1",
    "numpy==2.3.4",
    "faiss-cpu==1.12.0",
    "tiktoken==0.12.0",
    "rank-bm25==0.2.2",
    "pyyaml==6.0.3",
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []
only-include = [
    "*.py",
    "*.md",
    "*.txt"
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # 行の長さは100文字まで許可
